in ApPerformance change cpuUtilized type to int[], in ClientMetrics change mcs and rates types to int[] - they were originally defined as byte[], but that causes them to be serialized as base64-encoded strings, which is annoying

+++ all-cloud-and-opensync-gw-in-one-process/src/main/java/com/telecominfraproject/wlan/startuptasks/AllInOneWithGatewayStartListener.java
@@ -595,7 +595,7 @@ public class AllInOneWithGatewayStartListener implements ApplicationRunner {
 		apNodeMetrics.setChannelUtilization(RadioType.is5GHzU, getRandomInt(30, 70));
 
 		apPerformance.setCpuTemperature(getRandomInt(25, 90));
-		apPerformance.setCpuUtilized(new byte[] { (byte) getRandomInt(5, 98), (byte) getRandomInt(5, 98) });
+		apPerformance.setCpuUtilized(new int[] { getRandomInt(5, 98), getRandomInt(5, 98) });
 
 		apPerformance.setEthLinkState(EthernetLinkState.UP1000_FULL_DUPLEX);
